CON-107: Modified get_dictionary method to accept complex member_names
It is now possible to use the get_dictionary() method on an Input to obtain
a dictionary containing the fields of a nested member of a sample.
It is currently only possible to supply member_names of nested members with the
following kinds:
 - Array
 - Sequence
 - Struct
 - Value
 - Union
Updated docs and unit tests wrt to these changes.