CON-107: PR feedback
Minor documentation changes
Fixed a unit test that had not been updated to use 0-based indexes
Added a unit test to check that get_dictionary("foo") where foo is an unset
optional returns None. Since this is not yet implemented in the core I have
left it commented out for now.