CON-191: Updated tests and docs to reflect new behaviour
I noticed in the native libraries that we weren't being consistent with
the limits. Sometimes we were not checking for v < -LUA_MAX_INT, and sometimes
we were not being consistent with the use of > vs >=.
I have rectified this, the supported behaviour is as follows:

When setting via set_number, we allow |v| < 2^53.
When getting via get_number, we allow |v| <= 2^53.

The reason for the discrepancy between the set and get, is that
when we have a value in set_number, it has already been cast to a double,
so although 2^53 is perfectly representable, we do not know if the original
value was larger (e.g., 2^53 + 1 would appear as 2^53).

Also extended tests to ensure that all of the various supported ways
of setting / getting 64-bit values work as expected:
- allow setting large values via __setitem__ (string or number),
  set_string, set_dictionary
- allow getting large values via __getitem__ (returned as float or int
  depending on size), get_string, get_dictionary